pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bayesian-probabilistic-langchain-api"
version = "0.2.0"
description = "Portable FastAPI wrapper for Bayesian probabilistic LangChain orchestration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Angshul Majumdar"}]
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "pydantic>=2.6",
  "numpy>=1.24"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27"
]
gemini = [
  "google-genai>=1.0.0"
]
rag = [
  "sentence-transformers>=3.0.0",
  "transformers>=4.40.0",
  "torch>=2.2.0",
  "faiss-cpu>=1.8.0",
  "pandas>=2.2.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
